feat: route profile fixtures through direct checkers
