+++ packages/cosmo0/src/main/scala/cosmo0/tyck/README.md
@@ -8,8 +8,9 @@ patterns, mutability, and expression result types.
 
 `Profiles.scala` describes checker profiles and feature gates. `StageCapabilities.scala`
 validates package-level stage capability profiles before package checking.
-`ProfileCheckers.scala` adapts profile-specific source directives such as
-`mltt: lambda-checks-pi` into executable MLTT and dependent-pattern assertions.
+`MlttTypeChecker` owns `mltt.core` source directives such as
+`mltt: lambda-checks-pi`; `DependentPatterns` owns `mltt.dependent-patterns`
+source directives such as `dependent-pattern: vec-head-elaborates`.
 
 Subdirectories:
 
@@ -18,6 +19,7 @@ Subdirectories:
 
 Upstream input comes from `syntax/Elaborator.scala`. Downstream output goes to
 `lir/Lowerer.scala`. Profile metadata is also read by `source/Pipeline.scala`
-when package metadata selects a checker profile. `mltt.core` and
-`mltt.dependent-patterns` are routed through profile assertion checkers before
-`SourceTyper`; `cosmo0.subset` remains the ordinary source checker.
+when package metadata selects a checker profile. `mltt.core` is routed directly
+through `MlttTypeChecker`, while `mltt.dependent-patterns` is routed directly
+through `DependentPatterns`; `cosmo0.subset` remains the ordinary source
+checker.
