+++ packages/cosmo0/src/test/scala/cosmo0/CheckerProfileTests.scala
@@ -80,6 +80,22 @@ class CheckerProfileTests extends munit.FunSuite:
     assert(summary.contains("mltt_core_lambda_checks_pi"))
     assert(summary.contains("mltt_core_application_infers_through_pi"))
 
+  test("MLTT type checker exposes the cosmo0 source checker entry point"):
+    val source = SourceFile(
+      "mltt-direct.cos",
+      """mltt: lambda-checks-pi
+        |mltt: conversion-beta-let
+        |""".stripMargin,
+    )
+    val result = MlttTypeChecker.checkSource(source)
+
+    assertEquals(result.phase, Phase.Check)
+    assertEquals(result.status, PhaseStatus.Succeeded)
+    assert(result.diagnostics.isEmpty)
+    val summary = result.value.get.checkerArtifactSummary
+    assert(summary.contains("mltt_core_lambda_checks_pi"))
+    assert(summary.contains("mltt_core_conversion_beta_let"))
+
   test("MLTT profile rejects source that does not declare MLTT assertions"):
     val result = Cosmo0().checkWithProfile(
       """class Box {
@@ -198,6 +214,24 @@ class CheckerProfileTests extends munit.FunSuite:
     assertEquals(checked.phase, Phase.Check)
     assertEquals(checked.status, PhaseStatus.Succeeded)
     val summary = checked.value.get.checked.typed.checkerArtifactSummary
+    assert(summary.contains("dependent_pattern_vec_head_elaborates"))
+    assert(summary.contains("dependent_pattern_impossible_nil_diagnostic"))
+
+  test(
+    "dependent-pattern checker exposes the cosmo0 source checker entry point",
+  ):
+    val source = SourceFile(
+      "dependent-direct.cos",
+      """dependent-pattern: vec-head-elaborates
+        |dependent-pattern: impossible-nil-diagnostic
+        |""".stripMargin,
+    )
+    val result = DependentPatterns.checkSource(source)
+
+    assertEquals(result.phase, Phase.Check)
+    assertEquals(result.status, PhaseStatus.Succeeded)
+    assert(result.diagnostics.isEmpty)
+    val summary = result.value.get.checkerArtifactSummary
     assert(summary.contains("dependent_pattern_vec_head_elaborates"))
     assert(summary.contains("dependent_pattern_impossible_nil_diagnostic"))
